feat(modularization): extract schedule + notify into clawql-automation (phase 9)
Add packages/clawql-automation for persisted synthetic checks and Slack notify;
wire execute via configureNotifyDeps from tools.ts. Keep MCP log shaping in src shims.
