+++ docker/panguard-mcp-bridge/Dockerfile
@@ -17,6 +17,7 @@ COPY packages/clawql-core/package.json ./packages/clawql-core/package.json
 COPY packages/clawql-api/package.json ./packages/clawql-api/package.json
 COPY packages/clawql-memory/package.json ./packages/clawql-memory/package.json
 COPY packages/clawql-documents/package.json ./packages/clawql-documents/package.json
+COPY packages/clawql-automation/package.json ./packages/clawql-automation/package.json
 COPY packages/clawql-ouroboros/package.json ./packages/clawql-ouroboros/package.json
 COPY packages/mcp-grpc-transport ./packages/mcp-grpc-transport
 COPY packages/panguard-mcp-bridge ./packages/panguard-mcp-bridge
