+++ package.json
@@ -53,8 +53,8 @@
         ".cursor/mcp.json.example"
     ],
     "scripts": {
-        "clean": "rm -rf dist packages/clawql-core/dist packages/clawql-api/dist packages/clawql-memory/dist packages/clawql-documents/dist packages/mcp-grpc-transport/dist packages/clawql-ouroboros/dist packages/panguard-mcp-bridge/dist",
-        "build": "npm run clean && npm run build -w clawql-core && npm run build -w clawql-api && npm run build -w clawql-memory && npm run build -w clawql-documents && npm run build -w mcp-grpc-transport && npm run build -w clawql-ouroboros && npm run build -w panguard-mcp-bridge && tsc",
+        "clean": "rm -rf dist packages/clawql-core/dist packages/clawql-api/dist packages/clawql-memory/dist packages/clawql-documents/dist packages/clawql-automation/dist packages/mcp-grpc-transport/dist packages/clawql-ouroboros/dist packages/panguard-mcp-bridge/dist",
+        "build": "npm run clean && npm run build -w clawql-core && npm run build -w clawql-api && npm run build -w clawql-memory && npm run build -w clawql-documents && npm run build -w clawql-automation && npm run build -w mcp-grpc-transport && npm run build -w clawql-ouroboros && npm run build -w panguard-mcp-bridge && tsc",
         "build:turbo": "turbo run build",
         "prepublishOnly": "npm run build",
         "start": "node dist/server.js",
@@ -91,10 +91,10 @@
         "pretest": "npm run build",
         "test": "vitest run",
         "test:coverage": "npm run build && vitest run --coverage",
-        "lint": "eslint \"src/**/*.ts\" \"packages/clawql-core/src/**/*.ts\" \"packages/clawql-api/src/**/*.ts\" \"packages/clawql-memory/src/**/*.ts\" \"packages/clawql-documents/src/**/*.ts\" \"packages/mcp-grpc-transport/src/**/*.ts\" \"packages/clawql-ouroboros/src/**/*.ts\" \"packages/panguard-mcp-bridge/src/**/*.ts\" \"vitest.config.ts\" --max-warnings 0 --cache --cache-location .eslintcache",
-        "lint:fix": "eslint \"src/**/*.ts\" \"packages/clawql-core/src/**/*.ts\" \"packages/clawql-api/src/**/*.ts\" \"packages/clawql-memory/src/**/*.ts\" \"packages/clawql-documents/src/**/*.ts\" \"packages/mcp-grpc-transport/src/**/*.ts\" \"packages/clawql-ouroboros/src/**/*.ts\" \"packages/panguard-mcp-bridge/src/**/*.ts\" \"vitest.config.ts\" --fix --cache --cache-location .eslintcache",
-        "format": "prettier --write \"src/**/*.{ts,tsx}\" \"packages/clawql-core/**/*.{ts,md,json}\" \"packages/clawql-api/**/*.{ts,md,json}\" \"packages/clawql-memory/**/*.{ts,md,json}\" \"packages/clawql-documents/**/*.{ts,md,json}\" \"packages/mcp-grpc-transport/**/*.{ts,md,json}\" \"docs/**/*.md\" \".github/**/*.yml\" \"*.md\" \"vitest.config.ts\" \"eslint.config.js\" \"tsconfig.json\" \".prettierrc.json\" \"turbo.json\" --cache --cache-location .prettier-cache && npm run format --prefix website",
-        "format:check": "prettier --check \"src/**/*.{ts,tsx}\" \"packages/clawql-core/**/*.{ts,md,json}\" \"packages/clawql-api/**/*.{ts,md,json}\" \"packages/clawql-memory/**/*.{ts,md,json}\" \"packages/clawql-documents/**/*.{ts,md,json}\" \"packages/mcp-grpc-transport/**/*.{ts,md,json}\" \"docs/**/*.md\" \".github/**/*.yml\" \"*.md\" \"vitest.config.ts\" \"eslint.config.js\" \"tsconfig.json\" \".prettierrc.json\" \"turbo.json\" --cache --cache-location .prettier-cache && npm run format:check --prefix website",
+        "lint": "eslint \"src/**/*.ts\" \"packages/clawql-core/src/**/*.ts\" \"packages/clawql-api/src/**/*.ts\" \"packages/clawql-memory/src/**/*.ts\" \"packages/clawql-documents/src/**/*.ts\" \"packages/clawql-automation/src/**/*.ts\" \"packages/mcp-grpc-transport/src/**/*.ts\" \"packages/clawql-ouroboros/src/**/*.ts\" \"packages/panguard-mcp-bridge/src/**/*.ts\" \"vitest.config.ts\" --max-warnings 0 --cache --cache-location .eslintcache",
+        "lint:fix": "eslint \"src/**/*.ts\" \"packages/clawql-core/src/**/*.ts\" \"packages/clawql-api/src/**/*.ts\" \"packages/clawql-memory/src/**/*.ts\" \"packages/clawql-documents/src/**/*.ts\" \"packages/clawql-automation/src/**/*.ts\" \"packages/mcp-grpc-transport/src/**/*.ts\" \"packages/clawql-ouroboros/src/**/*.ts\" \"packages/panguard-mcp-bridge/src/**/*.ts\" \"vitest.config.ts\" --fix --cache --cache-location .eslintcache",
+        "format": "prettier --write \"src/**/*.{ts,tsx}\" \"packages/clawql-core/**/*.{ts,md,json}\" \"packages/clawql-api/**/*.{ts,md,json}\" \"packages/clawql-memory/**/*.{ts,md,json}\" \"packages/clawql-documents/**/*.{ts,md,json}\" \"packages/clawql-automation/**/*.{ts,md,json}\" \"packages/mcp-grpc-transport/**/*.{ts,md,json}\" \"docs/**/*.md\" \".github/**/*.yml\" \"*.md\" \"vitest.config.ts\" \"eslint.config.js\" \"tsconfig.json\" \".prettierrc.json\" \"turbo.json\" --cache --cache-location .prettier-cache && npm run format --prefix website",
+        "format:check": "prettier --check \"src/**/*.{ts,tsx}\" \"packages/clawql-core/**/*.{ts,md,json}\" \"packages/clawql-api/**/*.{ts,md,json}\" \"packages/clawql-memory/**/*.{ts,md,json}\" \"packages/clawql-documents/**/*.{ts,md,json}\" \"packages/clawql-automation/**/*.{ts,md,json}\" \"packages/mcp-grpc-transport/**/*.{ts,md,json}\" \"docs/**/*.md\" \".github/**/*.yml\" \"*.md\" \"vitest.config.ts\" \"eslint.config.js\" \"tsconfig.json\" \".prettierrc.json\" \"turbo.json\" --cache --cache-location .prettier-cache && npm run format:check --prefix website",
         "release:git-tag": "bash scripts/release/git-release-tag.sh",
         "release:git-tag:push": "bash scripts/release/git-release-tag.sh --push",
         "release:git-backfill-tags": "bash scripts/release/git-backfill-npm-release-tags.sh",
@@ -113,6 +113,7 @@
         "@opentelemetry/sdk-trace-node": "^2.7.1",
         "clawql-api": "workspace:*",
         "clawql-core": "workspace:*",
+        "clawql-automation": "file:../packages/clawql-automation",
         "clawql-documents": "file:../packages/clawql-documents",
         "clawql-memory": "file:../packages/clawql-memory",
         "clawql-ouroboros": "workspace:*",
